refactor(zones): ♻️ Move `GetZoneCoordinatesFromGPS` from the utils folder
This was needed to avoid the Cyclic Impor error
